pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-format-converter"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["PySide6>=6.8", "Pillow>=11.0"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-qt>=4.4", "pyinstaller>=6.0"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"
